move solution back to helper
